#15 検索結果リストへ検索結果をリスト化した - routerだと配列を値として遷移先へ渡すのが難しかったので、検索ボックス内にロジックを書かず、検索結果ページにロジックを移した
