+++ src/components/Top/SearchBox.tsx
@@ -1,14 +1,9 @@
 import { useRouter } from 'next/router';
-import React, { useCallback, useEffect, useState } from 'react'
-import { useRecoilState } from 'recoil';
+import React, { useCallback, useState } from 'react'
 import { SearchBoxLayouts } from '../Layouts/SearchBoxLayouts';
-import { WordsAtom } from '../utils/atoms/WordsAtom';
-import type { Words } from '../types/types';
 
 export const SearchBox = () => {
-    const [ words, setWords] = useRecoilState(WordsAtom)
     const [ keyWord, setKeyWord ] = useState('')
-    const [ result, setResult ]= useState([])
     const router = useRouter()
     const handleChange = useCallback((e) => {
         setKeyWord(e.target.value)
@@ -16,39 +11,14 @@ export const SearchBox = () => {
 
     const handleClickSearch = () => {
 
-        //KeyWordが入力されたら
-        if(keyWord) {
-            const filteredWords = words.filter((item:Words) => {
-                    return (
-                      item.title.toLowerCase().indexOf(keyWord) !== -1 ||
-                      item.description.toLowerCase().indexOf(keyWord) !== -1 ||
-                      item.shortParaphrase.toLowerCase().indexOf(keyWord) !== -1
-                    );
-            });
-
-             setResult([...result].concat(filteredWords))
-        }
-
-        //mapで回して1つずつ
-        // result.map(item =>
-        //   {
-        //     return [item.title, item.description, item.shortParaphrase]
-        // })
-
-        //このonClickで発動する関数内だと空配列
-        console.log(result);
-
         router.push({
 
-          // pathname:"/SearchResult/SearchResultList/",
+          pathname:"/SearchResult/SearchResultList/",
           query: {
-           
+           keyWord
           }
         })
-
       }
-      //関数の外なら空ではない
-      console.log("result 2",result)
 
   return (
     <div>
